requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
SQLAlchemy==2.0.40
alembic==1.15.2
pydantic==2.11.1
pydantic[email]
pytest
httpx==0.28.1
python-jose[cryptography] 
passlib[bcrypt]